fixed some bugs in settings
there are still problems with configuring children though

- The place holder of a child's birthday becomes yyyy-mm-dd
which is not the format we want, and there is no code that checks user
input for b-day. If an incorrect format is used, this breaks every page
that uses the childJsonObject localStorage object.

- Even though we enter the right format when editing children, for some
reason another child with a wackjob b-day is formed.. new child entry is
entered into local storage